requirements.txt
Flask==0.12.2
gevent==1.2.2
ConcurrentLogHandler==0.9.1
rsa==3.4.2
flask-restplus==0.10.1
Flask-SQLAlchemy==2.2
mysql-python==1.2.5
enum==0.4.6
flask-migrate==2.1.1
flask-script==2.0.6
axf-library==1.0.24
pycrypto==2.6.1
pyOpenSSL==17.5.0
requests==2.18.1
celery==4.0.2
kombu==4.1.0
attrs==18.1.0
paramiko==2.4.1
redis==2.10.5